requirements.txt
python-dotenv
discord.py
asyncio
python-dateutil
requests
pylint
